benchmark_vortex_parquet/requirements.txt
vortex-data==0.58.0
parquet==1.3.1
requests>=2.31.0
pyarrow==23.0.0
duckdb==1.4.3
polars>=1.0.0
numpy>=2.4.1
matplotlib>=3.7.0
pandas>=2.0.0